+++ engine/include/system/system.h
@@ -63,7 +63,7 @@ extern uint8_t sys_dump_sreg(const struct _system *this);
 
 /* FLASH Operations */
 
-extern progmem_t* sys_read_instr(const struct _system *this, const int addr);
+extern progmem_t* sys_read_progmem(const struct _system *this, const int addr);
 extern uint16_t sys_read_flash(const struct _system *this, const int addr);
 extern void sys_move_pc(const struct _system *this, const int inc);
 extern void sys_set_pc(struct _system *this, const int addr);